Parameterized functions
Defines a general framework (based on the OOP-capabilities in MATLAB) to
define a set of parameterized functions. It allows an arbitrary number
of user-specified parameters to be associated with any number of
user-defined functions. The associated functions within the function set
is automatically evaluated and the result array updated whenever a
parameter value is changed.